Use a PID file to track starting/stopping daemons.

+++ launcher/archivesspace.sh
@@ -40,25 +40,46 @@ startup_cmd="java "$JAVA_OPTS"  \
         org.jruby.Main --1.9 \"launcher/launcher.rb\""
 
 
+export PIDFILE=$ASPACE_LAUNCHER_BASE/.archivesspace.pid
+
+
 case "$1" in
     start)
+        if [ -e "$PIDFILE" ]; then
+            pid=`cat $PIDFILE 2>/dev/null`
+
+            if [ "$pid" != "" ] && kill -0 $pid &>/dev/null; then
+                echo "There already seems to be an instance running (PID: $pid)"
+                exit
+            fi
+        fi
+
         shellcmd="bash"
         if [ "$ARCHIVESSPACE_USER" != "" ]; then
             shellcmd="su - $ARCHIVESSPACE_USER"
         fi
 
         $shellcmd -c "cd '$ASPACE_LAUNCHER_BASE';
-          (exec 0<&-; exec 1>&-; exec 2>&-; $startup_cmd &> \"logs/archivesspace.out\" & ) &
+          (
+             exec 0<&-; exec 1>&-; exec 2>&-;
+             $startup_cmd &> \"logs/archivesspace.out\" &
+             echo \$! > \"$PIDFILE\"
+          ) &
           disown $!"
 
         echo "ArchivesSpace started!  See logs/archivesspace.out for details."
         ;;
     stop)
-        pid=`ps -ef | grep 'Darchivesspace-daemon=yes' | grep -v grep | awk '{print $2}' | head -1`
+        pid=`cat $PIDFILE 2>/dev/null`
         if [ "$pid" != "" ]; then
-            echo -n "Shutting down ArchivesSpace (running as PID $pid)... "
-            kill $pid
-            echo "done"
+            kill -0 $pid &>/dev/null
+            if [ "$?" = "0" ]; then
+                echo -n "Shutting down ArchivesSpace (running as PID $pid)... "
+                kill $pid
+                echo "done"
+            fi
+
+            rm -f "$PIDFILE"
         else
             echo "Couldn't find a running instance to stop"
         fi
